:recycle: (video): Set CGGraphics with CGColor+CGPoint+CGPixel+Character+FilledRectangle
