+++ drivers/CoreVideo/tests/CGColor_test.cpp
@@ -2,8 +2,7 @@
 // Copyright 2021 APF France handicap
 // SPDX-License-Identifier: Apache-2.0
 
-#include "CGColor.hpp"
-
+#include "CGGraphics.hpp"
 #include "gtest/gtest.h"
 
 using namespace leka;
