+++ drivers/CoreVideo/tests/CoreVideo_test.cpp
@@ -4,7 +4,6 @@
 
 #include "CoreVideo.hpp"
 
-#include "Character.hpp"
 #include "gtest/gtest.h"
 #include "mocks/leka/CoreDMA2D.h"
 #include "mocks/leka/CoreDSI.h"
